refactor: transfergraphic component
